IEP-1584 Avoid Simultaneous Modals on Old Config Detection (#1264)
fix: added global modal lock to avoid multiple modals

+++ bundles/com.espressif.idf.ui/src/com/espressif/idf/ui/LaunchBarListener.java
@@ -140,15 +140,17 @@ public class LaunchBarListener implements ILaunchBarListener
 						// If both are not same
 						if (currentTarget != null && !newTarget.equals(currentTarget))
 						{
-
-							boolean isDelete = MessageDialog.openQuestion(EclipseUtil.getShell(),
+							GlobalModalLock.showModal(() -> MessageDialog.openQuestion(EclipseUtil.getShell(),
 									Messages.LaunchBarListener_TargetChanged_Title,
 									MessageFormat.format(Messages.LaunchBarListener_TargetChanged_Msg,
-											project.getName(), currentTarget, newTarget));
-							if (isDelete)
-							{
-								deleteBuildFolder(project, buildLocation);
-							}
+											project.getName(), currentTarget, newTarget)),
+									isDelete -> {
+										if (isDelete)
+										{
+											deleteBuildFolder(project, buildLocation);
+
+										}
+									});
 						}
 					}
 				}
